Hash team passwords
Move hashing function from `user` module to `common`
TODO: Update all team tests

IMPORTANT: If you have existing teams created in plaintext, the
following script will hash all plaintext team passwords and is
safe to run more than once. This all applies to the `web` instance.

First, set environment variable:
`APP_SETTINGS_FILE=/picoCTF-web-config/deploy_settings.py`

Then execute this script within the `/picoCTF-env` virtualenv:

```
import api

all_teams = api.team.get_all_teams(show_ineligible=True)
for team in all_teams:
    if isinstance(team["password"], str):
            api.team.update_password(team["tid"], team["password"])

```
